Revert "fix: :bug: Delete verification after identity set"
